requirements.txt
tensorflow-cpu
numpy
opencv-python-headless
pandas
matplotlib
scikit-learn
gunicorn
deep-translator
ultralytics
easyocr
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
Pillow
